LayoutTests/fast/dom/shadow/shadow-dom-event-dispatching-child-of-inactive-content.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../js/resources/js-test-pre.js"></script>
    <script src="resources/shadow-dom.js"></script>
    <script src="resources/event-dispatching.js"></script>
</head>
<body>
    <p id="description"></p>
    <div id="sandbox"></div>
    <pre id="console"></pre>
    <script>
        var sandbox = document.getElementById('sandbox');

        sandbox.appendChild(
            createDOM('div', {'id': 'top'},
                      createDOM('div', {'id': 'A'}),
                      createDOM('div', {'id': 'B'},
                                createShadowRoot(
                                    createDOM('content', {'id': 'active-content', 'select': '#parent-of-inactive-content'})),
                                createDOM('div', {'id': 'parent-of-inactive-content'},
                                          createDOM('content', {'id': 'inactive-content'},
                                                    createDOM('div', {'id': 'child-of-inactive-content'}))))));

        addEventListeners(['top', 'A', 'B', 'B/', 'B/active-content',
                           'parent-of-inactive-content', 'inactive-content', 'child-of-inactive-content']);
        getNodeInShadowTreeStack('B/').id = 'shadow-root-B';
        showSandboxTree();
        moveMouse('A', 'child-of-inactive-content');
    </script>
    <script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
